requirements.txt
Django==5.2.*
psycopg[binary]==3.2.*
djangorestframework==3.15.*
django-environ==0.11.*
django-cors-headers==4.4.*
django-filter==24.2
gunicorn==22.0.*
djangorestframework-simplejwt==5.3.*
drf-spectacular==0.26.*
django-extensions==3.2.*
python-decouple==3.8
